Refactor code to separate RealtimeBlackboard as its own component
