+++ src/pages/ClassPageLiveBoard.vue
@@ -2,72 +2,44 @@
   <div id="room">
     <template v-if="user">
       <LiveBoardAudio :roomId="roomId"/>
-      <p v-if="!hasFetchedStrokesFromDb">Loading the real-time blackboard...</p>
-      <Blackboard v-else
-        ref="Blackboard"
-        isRealtime
-        :strokesArray="strokesArray"
-        @stroke-drawn="stroke => uploadToDb(stroke)"
-        @board-reset="deleteAllStrokesFromDb()"
-      >
+      <RealtimeBlackboard :strokesRef="strokesRef"> 
         <template v-slot:blackboard-toolbar>
-          <ButtonNew @click="toggleHelpSignal()" icon="mdi-account-alert">Signal For Help</ButtonNew> 
-          <ButtonNew @click="uploadExplanation()" icon="mdi-upload">Save Board</ButtonNew>
-        </template>
-      </Blackboard> 
+          <ButtonNew @click="toggleHelpSignal()" icon="mdi-account-alert">
+            Signal For Help
+          </ButtonNew> 
+        </template> 
+      </RealtimeBlackboard>
     </template>
   </div>
 </template>
 
 <script>
-/**
- * A real-time blackboard, which is <Blackboard/> hooked up to Firestore.
- * 
- * Proof of database <=> blackboard: 
- * 
- * database => blackboard: 
- *     Whenever a new stroke is added to the database, the database listener will trigger and update 
- *     strokesArray, which propagates to the Blackboard component
- * 
- * blackboard => database: 
- *     Whenever the user draws a new stroke, that stroke will be uploaded as a new document to the database. 
- *     Momentarily, there will be divergence between state and UI:
- *     the user's `strokesArray` does NOT have the new stroke, even though she was the one who drew the stroke.
- *     However, when the write operation finally resolves, everyone's `strokesArray` will be updated, 
- *     including the original author herself. Note that there will be double drawing, where the first is for feedback,
- *     and the second is to signal when other people can actually see her new stroke. 
- */
 import firebase from "firebase/app";
 import "firebase/firestore";
-import Blackboard from "@/components/Blackboard.vue";
 import LiveBoardAudio from "@/components/LiveBoardAudio.vue";
 import DatabaseHelpersMixin from "@/mixins/DatabaseHelpersMixin.js";
-import ExplUploadHelpers from "@/mixins/ExplUploadHelpers.js";
 import db from "@/database.js";
 import ButtonNew from "@/components/ButtonNew.vue";
 import { mapState } from "vuex";
-import { getRandomId } from "@/helpers.js";
+import RealtimeBlackboard from "@/components/RealtimeBlackboard.vue";
 
 export default {
   components: { 
     ButtonNew,
-    Blackboard,
     LiveBoardAudio,
+    RealtimeBlackboard
   },
   mixins: [
     DatabaseHelpersMixin,
-    ExplUploadHelpers
   ],
   data () {
     return {
-      strokesArray: [],
       room: {},
       roomRef: null,
       strokesRef: null,
       unsubscribeRoomListener: null,
       classId: this.$route.params.class_id,
       roomId: this.$route.params.room_id,
-      hasFetchedStrokesFromDb: false
     }
   },
   computed: {
@@ -88,15 +60,11 @@ export default {
   async created () {
     this.roomRef = db.doc(`classes/${this.classId}/blackboards/${this.roomId}`);
     this.strokesRef = this.roomRef.collection("strokes");
-    // this.room = await this.$_getDoc(this.roomRef);
     this.unsubscribeRoomListener = await this.$_listenToDoc(this.roomRef, this, "room"); 
     this.setUserDisconnectHook();
-    this.keepSyncingBoardWithDb();
   },
   beforeDestroy () {
-    if (this.unsubscribeRoomListener) {
-      this.unsubscribeRoomListener();
-    }
+    this.unsubscribeRoomListener();
     this.roomRef.update({
       participants: firebase.firestore.FieldValue.arrayRemove(this.simplifiedUser)
     });
@@ -106,79 +74,6 @@ export default {
       await this.roomRef.update({
         status: this.room.status === "Needs help" ? "" : "Needs help"
       }); 
-    },
-    async uploadExplanation () {
-      const { TextEditor, Blackboard } = this.$refs;
-      const title = `Untitled (${new Date().toLocaleTimeString()})`; 
-      const html = "";
-      const explRef = db.doc(`classes/${this.mitClass.id}/posts/${getRandomId()}`);
-      const audioBlob = null; // TODO
-      const thumbnailBlob = await Blackboard.getThumbnail();
-      this.$_saveExplToCacheThenUpload(
-        thumbnailBlob,
-        audioBlob,
-        html,
-        title,
-        explRef
-      );
-    },
-    /*
-      TODO:
-        1. order by `timestamp` rather than `strokeNumber`
-        2. Investigate why each new stroke triggers 2-3 snapshot callbacks
-        3. handle a board reset more efficiently
-    */
-    keepSyncingBoardWithDb () {
-      this.strokesRef.orderBy("strokeNumber").onSnapshot(snapshot => {
-        if (snapshot.docs.length === 0) {
-          this.strokesArray = [];
-        } else {
-          snapshot.docChanges().filter(change => change.type === "added").forEach(change => {
-            this.strokesArray.push({
-              id: change.doc.id,
-              ...change.doc.data()
-            });
-          });
-        }
-
-        if (!this.hasFetchedStrokesFromDb) { 
-          this.hasFetchedStrokesFromDb = true;
-        }
-      });
-    },
-    async uploadToDb (stroke) {
-      const timestamp = firebase.firestore.FieldValue.serverTimestamp();
-      try {
-        this.strokesRef.add({
-          timestamp,
-          ...stroke
-        });
-      } catch (error) {
-        this.$root.$emit("snow-snackbar", "Failed to upload stroke to database.");
-      }
-    },
-    async deleteAllStrokesFromDb () {
-      const promises = [];
-      const strokeDeleteRequests = [];
-      let currentBatch = db.batch();
-      let currentBatchSize = 0;
-
-      for (const stroke of this.strokesArray) {
-        if (currentBatchSize >= 500) {
-          promises.push(currentBatch.commit());
-          currentBatch = db.batch(); 
-          currentBatchSize = 0; 
-        } 
-        currentBatch.delete(this.strokesRef.doc(stroke.id));
-        currentBatchSize += 1;
-      }
-
-      promises.push(currentBatch.commit()); 
-      promises.push(
-        this.roomRef.update({ imageUrl: "" })
-      );
-      await Promise.all(promises);
-      this.$root.$emit("show-snackbar", "Successfully reset blackboard.");
     },
     setUserDisconnectHook () {
       /*
